array_to_str in utils
